requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
easyocr==1.7.1
exceptiongroup==1.2.0
fastapi==0.110.1
fasteners==0.14.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
future==0.18.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
imageio==2.34.0
importlib-metadata==4.6.4
imutils==0.5.4
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
paramiko==2.9.3
pdf2image==1.17.0
pexpect==4.8.0
pillow==10.3.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pyclipper==1.3.0.post5
pycups==2.0.1
pydantic==2.7.0
pydantic_core==2.18.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.9
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scikit-image==0.23.1
scikit-learn==1.4.1.post1
scipy==1.13.0
SecretStorage==3.3.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
Sublist3r==1.0
sympy==1.12
systemd-python==234
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.29.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
